+++ Packages/Tracking Preview/UI Input/Runtime/Scripts/PointerElement.cs
@@ -29,6 +29,7 @@ namespace Leap.InputModule
         [SerializeField] private UIInputCursor cursor;
         [SerializeField] private bool forceDisable = false;
         [SerializeField] private bool disableWhenOffCanvas = true;
+        [SerializeField] private Finger.FingerType finger = Finger.FingerType.INDEX;
 
         public Chirality Chirality { get; private set; }
 
@@ -153,13 +154,11 @@ namespace Leap.InputModule
         }
 
         /// <summary>
-        /// The z position of the index finger tip to the Pointer
+        /// The z position of the finger tip to the Pointer
         /// </summary>
-        private float DistanceOfTipToPointer(Hand hand)
+        public float DistanceOfTipToPointer(Hand hand)
         {
-            var tipPosition = hand.fingers[(int)Finger.FingerType.INDEX]
-                .GetBone(Bone.BoneType.DISTAL).NextJoint;
-
+            var tipPosition = hand.fingers[(int)finger].GetBone(Bone.BoneType.DISTAL).NextJoint;
             var pointerTransform = transform;
             return -pointerTransform.transform.InverseTransformPoint(tipPosition).z * pointerTransform.transform.lossyScale.z - module.TactilePadding;
         }
@@ -474,7 +473,7 @@ namespace Leap.InputModule
 
         private void ProcessTactile(IProjectionOriginProvider projectionOriginProvider, Hand hand)
         {
-            // Raycast from shoulder through tip of the index finger to the UI
+            // Raycast from shoulder through tip of the finger to the UI
             var tipRaycastUsed = GetLookPointerEventData(
                 hand,
                 projectionOriginProvider.ProjectionOriginForHand(hand),
@@ -752,24 +751,7 @@ namespace Leap.InputModule
             if (IsTouchingOrNearlyTouchingCanvasOrElement() || forceTipRaycast)
             {
                 tipRaycast = true;
-
-                var farthest = 0f;
-                pointerPosition = hand.Index.TipPosition;
-                for (var i = 1; i < 3; i++)
-                {
-                    var fingerDistance = Vector3.Distance(module.MainCamera.transform.position,
-                        hand.fingers[i].TipPosition);
-                    var fingerExtension =
-                        Mathf.Clamp01(Vector3.Dot(
-                            hand.fingers[i].Direction,
-                            module.LeapDataProvider.CurrentFrame.Hands[0].Direction)) / 1.5f;
-
-                    if (fingerDistance > farthest && fingerExtension > 0.5f)
-                    {
-                        farthest = fingerDistance;
-                        pointerPosition = hand.fingers[i].TipPosition;
-                    }
-                }
+                pointerPosition = hand.fingers[(int)Finger.FingerType.INDEX].TipPosition;
             }
             else
             {