+++ Packages/Tracking Preview/UI Input/Runtime/Scripts/UIInputCursor.cs
@@ -16,7 +16,13 @@ namespace Leap.InputModule
     public class UIInputCursor : MonoBehaviour
     {
         [SerializeField] private PointerElement element;
-        [SerializeField] private float interactionPointerScale = 0.6f;
+
+        [Tooltip("The minimum scale of the pointer when reacting to the user's interaction proximity (either pinch or distance, depending on mode)")]
+        [Range(0, 1)] [SerializeField] private float interactionPointerScale = 0.4f;
+        [Tooltip("The range of user interaction distance from the canvas at which to scale the pointer, when in direct mode")]
+        [Range(0, 1)][SerializeField] private float interactionPointerRange = 0.2f;
+        [Tooltip("If directly interacting, this will fade the cursor at the distance range specified above")]
+        [SerializeField] private bool fadeCursorAtDistance = true;
 
         private SpriteRenderer spriteRenderer;
         private Vector3 initialScale;
@@ -61,17 +67,26 @@ namespace Leap.InputModule
                 spriteRenderer.enabled = true;
             }
 
-            spriteRenderer.transform.localScale = hand != null
-                ? Vector3.Lerp(initialScale, initialScale * interactionPointerScale, hand.PinchStrength)
-                : Vector3.one;
+            if (element.IsUserInteractingDirectly)
+            {
+                spriteRenderer.transform.localScale = hand != null
+                    ? Vector3.Lerp(initialScale * interactionPointerScale, initialScale, element.DistanceOfTipToPointer(hand).Map(0.0f, interactionPointerRange, 0.0f, 1.0f))
+                    : Vector3.one;
+            }
+            else
+            {
+                spriteRenderer.transform.localScale = hand != null
+                    ? Vector3.Lerp(initialScale, initialScale * interactionPointerScale, hand.PinchStrength)
+                    : Vector3.one;
+            }
 
             switch (element.AggregatePointerState)
             {
-                case PointerStates.OnCanvas:
-                    spriteRenderer.color = colorBlock.normalColor;
-                    break;
                 case PointerStates.OffCanvas:
                     spriteRenderer.color = colorBlock.disabledColor;
+                    return;
+                case PointerStates.OnCanvas:
+                    spriteRenderer.color = colorBlock.normalColor;
                     break;
                 case PointerStates.OnElement:
                     spriteRenderer.color = colorBlock.highlightedColor;
@@ -94,6 +109,19 @@ namespace Leap.InputModule
                 default:
                     spriteRenderer.color = colorBlock.normalColor;
                     break;
+            }
+
+            if (element.IsUserInteractingDirectly && fadeCursorAtDistance)
+            {
+                spriteRenderer.color = new Color(
+                    spriteRenderer.color.r, 
+                    spriteRenderer.color.g, 
+                    spriteRenderer.color.b, 
+                    spriteRenderer.transform.localScale.x.Map(
+                        initialScale.x, 
+                        initialScale.x * interactionPointerScale, 
+                        0.0f,
+                        1.0f));
             }
         }
     }